fix bug in direntry mode conversion
Dir.mode is a u32
Sys_dir.mode is a WORD

When converting a ~0 value from Dir.mode to Sys_dir.mode, it is not
sign extended.

The default value for Dir.mode (specified by nulldir) is ~0.

When checking for mode's default value in Sys_dir, it should be checked
for 16rFFFFFFFF along with ~0.